requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
typer>=0.9.0
openai>=1.0.0
anthropic>=0.7.0
Pillow>=10.0.0